% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{is_in}
\alias{is_in}
\title{wraps %in% function for use in pipe useful}
\usage{
is_in(col1, col2, top = FALSE)
}
\arguments{
\item{col1}{col1}

\item{col2}{col2}

\item{top}{top}
}
\value{

}
\description{
wraps %in% function for use in pipe useful
}
