% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{model_analysis}
\alias{model_analysis}
\title{Title}
\usage{
model_analysis(xw, sw_input, sw_results, n = 3)
}
\arguments{
\item{xw}{The crosswalk}

\item{sw_input}{the input for SOCWalk}

\item{sw_results}{The socwalk results}

\item{n}{the number of top responses from socwalk to maintain}
}
\value{
a tibble of socwalk results..
}
\description{
Title
}
