% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{nest_results}
\alias{nest_results}
\title{Nest SOCwalk results}
\usage{
nest_results(data, cols, new_col_name)
}
\arguments{
\item{data}{the data}

\item{cols}{the cols you want to nest}

\item{new_col_name}{the col name for the nested data}
}
\value{
a tibble with the data nested
}
\description{
Nest SOCwalk results
}
