% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{soc_agree_5d}
\alias{soc_agree_5d}
\alias{soc_agree_3d}
\alias{soc_agree_2d}
\title{get the n-digit level of agreement...}
\usage{
soc_agree_5d(col1, col2, top = FALSE)

soc_agree_3d(col1, col2, top = FALSE)

soc_agree_2d(col1, col2, top = FALSE)
}
\arguments{
\item{col1}{String -- The column title of SOCWalk codes}

\item{col2}{String -- The column of expert assigned codes}

\item{top}{logical -- Should we only include the top SOCWalk code (default=FALSE)}
}
\value{
rv
}
\description{
get the n-digit level of agreement...
}
\section{Functions}{
\itemize{
\item \code{soc_agree_3d}: soc agree 3d

\item \code{soc_agree_2d}: soc agree 2d
}}

