% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{top_n_results}
\alias{top_n_results}
\title{returns the top n results from a nested data frame}
\usage{
top_n_results(data, col, codes_to, score_to, n = 3)
}
\arguments{
\item{data}{data}

\item{col}{the column}

\item{codes_to}{ccc}

\item{score_to}{ss}

\item{n}{n}
}
\value{

}
\description{
returns the top n results from a nested data frame
}
